corrections to main and functions

+++ ctof.h
@@ -1,1 +1,1 @@
-int ctof(float c);
+float ctof(float c);
